Improved mm, mutex and other tiny fixes.

+++ local.h
@@ -3,9 +3,7 @@
 
 namespace __cxxabiv1
 {
-	//_extension__ typedef int __guard __attribute__((mode(__DI__)));
-	typedef int __guard __attribute__((mode(__DI__)));
-	//typedef __int64_t __guard;
+	typedef long  __guard __attribute__((mode(__DI__)));
 
 	extern "C" {
 		int __cxa_guard_acquire(__guard *);
